Menu class and button titles
Added functionality to Menu class to create and display buttons.
Menus can set button positions, and bind their callbacks on creation.
Modified main function to create a new menu with 3 difficulty levels to start.

Added bShouldDestroy bool for deffered destruction in the Scene class. This is used due to the Menu class iterating over buttons and still trying to access the data when it was destroyed.
CleanUpMarkedScenes now deletes all Scene objects from the sceneStack at the end of a frame.