End game screens
Added end game menu and win lose displays.
Made some changes to the Scene class to get the size of a scene.
Added function to get centered position of the scene.
Added Initialize function for post construction loading of resources and other tasks.
Made Clear Resources public to remove it from the destructor.
In the future could probably make a resource or scene manager class so these aren't public